Add Tabs to Metrics Page
The metrics page which shows the overall cluster metrics, now contains
tabs (only the overview tab is implemented for now) to show different
information similar to the other pages.

Besides that we also added a nodes section to the overview tab, which
shows the nodes in the cluster with their cpu and memory information.
